chore: separate files for HMR demo
