+++ examples/react-router-ssr-data-router/src/entry.client.tsx
@@ -6,7 +6,7 @@ import {
   RouterProvider,
 } from "react-router-dom";
 
-import { routes } from "./App";
+import { routes } from "./app";
 
 hydrate();
