+++ examples/react-router-ssr-data-router/src/home.loader.ts
@@ -1,0 +1,6 @@
+import { rand, sleep } from "./utils";
+
+export async function homeLoader() {
+  await sleep();
+  return { data: `Home loader - random value ${rand()}` };
+}
