+++ examples/react-router-ssr-data-router/src/home.tsx
@@ -1,0 +1,12 @@
+import { useLoaderData } from "react-router-dom";
+import type { homeLoader } from "./home.loader";
+
+export function Home() {
+  let data = useLoaderData() as Awaited<ReturnType<typeof homeLoader>>;
+  return (
+    <div>
+      <h2>Home</h2>
+      <p>Loader Data: {data.data}</p>
+    </div>
+  );
+}
