+++ examples/react-router-ssr-data-router/src/utils.ts
@@ -1,0 +1,2 @@
+export const sleep = (n = 500) => new Promise((r) => setTimeout(r, n));
+export const rand = () => Math.round(Math.random() * 100);
